calling get_plaidml_device out of main process to avoid crash with plaidml
